refactor: database to typeorm
