+++ src/shared/infra/http/server.ts
@@ -4,7 +4,7 @@ import 'dotenv/config'
 import swaggerUi from 'swagger-ui-express'
 import swaggerDocument from '../../../swagger.json'
 
-import './database'
+import './shared/infra/typeorm'
 
 import './shared/container'
